+++ cgm/core/lib/rayCaster.py
@@ -51,7 +51,7 @@ log.setLevel(logging.INFO)
 def get_eligibleMesh():
     _str_func = ''
     _res = []
-    for l in mc.ls(type='mesh',visible = True, int=True), mc.ls(type='nurbsSurface',int=True, visible = True):
+    for l in mc.ls(type='mesh',visible = True, long=True), mc.ls(type='nurbsSurface',long=True, visible = True):
         for o in l:
             _mesh =  SHAPE.get_nonintermediate(o)
             _type = VALID.get_mayaType(_mesh)
