+++ cgm/lib/classes/AttrFactory.py
@@ -834,7 +834,7 @@ class AttrFactory():
         success(bool)
         """
         assert mc.objExists(target),"'%s' doesn't exist"%target
-        assert mc.ls(target,int=True) != [self.obj.nameShort], "Can't transfer to self!"
+        assert mc.ls(target,long=True) != [self.obj.nameShort], "Can't transfer to self!"
         functionName = 'doCopyTo'
         
         convertToMatch = kw.pop('convertToMatch',True)
@@ -907,7 +907,7 @@ class AttrFactory():
         assert mc.objExists(target),"'%s' doesn't exist"%target
         assert mc.ls(target,type = 'transform',int = True),"'%s' Doesn't have a transform"%target
         assert self.obj.transform is not False,"'%s' Doesn't have a transform. Transferring this attribute is probably a bad idea. Might we suggest doCopyTo along with a connect to source option"%self.obj.nameShort        
-        assert mc.ls(target,int=True) != [self.obj.nameShort], "Can't transfer to self!"
+        assert mc.ls(target,long=True) != [self.obj.nameShort], "Can't transfer to self!"
         assert '.' not in list(target),"'%s' appears to be an attribute. Can't transfer to an attribute."%target
         assert self.dynamic is True,"'%s' is not a dynamic attribute."%self.nameCombined
